fix: validate classic protocol parameter updates

+++ ledger/alonzo/pparams.go
@@ -68,7 +68,7 @@ type AlonzoProtocolParameters struct {
 	MaxBlockHeaderSize   uint
 	KeyDeposit           uint
 	PoolDeposit          uint
-	MaxEpoch             uint
+	MaxEpoch             uint64
 	NOpt                 uint
 	A0                   *cbor.Rat
 	Rho                  *cbor.Rat
@@ -280,7 +280,7 @@ type AlonzoProtocolParameterUpdate struct {
 	MaxBlockHeaderSize   *uint                                     `cbor:"4,keyasint"`
 	KeyDeposit           *uint                                     `cbor:"5,keyasint"`
 	PoolDeposit          *uint                                     `cbor:"6,keyasint"`
-	MaxEpoch             *uint                                     `cbor:"7,keyasint"`
+	MaxEpoch             *uint64                                   `cbor:"7,keyasint"`
 	NOpt                 *uint                                     `cbor:"8,keyasint"`
 	A0                   *cbor.Rat                                 `cbor:"9,keyasint"`
 	Rho                  *cbor.Rat                                 `cbor:"10,keyasint"`
