+++ ledger/babbage/pparams.go
@@ -36,7 +36,7 @@ type BabbageProtocolParameters struct {
 	MaxBlockHeaderSize   uint
 	KeyDeposit           uint
 	PoolDeposit          uint
-	MaxEpoch             uint
+	MaxEpoch             uint64
 	NOpt                 uint
 	A0                   *cbor.Rat
 	Rho                  *cbor.Rat
@@ -158,7 +158,7 @@ type BabbageProtocolParameterUpdate struct {
 	MaxBlockHeaderSize   *uint                                     `cbor:"4,keyasint"`
 	KeyDeposit           *uint                                     `cbor:"5,keyasint"`
 	PoolDeposit          *uint                                     `cbor:"6,keyasint"`
-	MaxEpoch             *uint                                     `cbor:"7,keyasint"`
+	MaxEpoch             *uint64                                   `cbor:"7,keyasint"`
 	NOpt                 *uint                                     `cbor:"8,keyasint"`
 	A0                   *cbor.Rat                                 `cbor:"9,keyasint"`
 	Rho                  *cbor.Rat                                 `cbor:"10,keyasint"`
